f(pid_for_speed_control): Customize pwm values for mbed pwm driver

+++ ports/mbed_common/include/drivers/actuators/dc_motor/mbed_dc_motor_ft_smc.h
@@ -28,7 +28,7 @@ public:
     /**
     * @brief Validate the set motor speed value
     */
-    unsigned int validate_motor_speed(unsigned int speed);
+    float validate_motor_speed(unsigned int speed);
 
     /**
     * @brief Sets a motor's raw speed
